+++ app/src/main/java/com/example/ktunsplashapp/presentation/MainActivity.kt
@@ -1,0 +1,41 @@
+package com.example.ktunsplashapp.presentation
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.ui.Modifier
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import com.example.ktunsplashapp.presentation.ui.theme.KtUnsplashAppTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            KtUnsplashAppTheme {
+                // A surface container using the 'background' color from the theme
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colorScheme.background
+                ) {
+                    val navController = rememberNavController()
+                    NavHost(
+                        navController = navController,
+                        startDestination = ScreenRoutes.SearchPhotosScreen.route
+                    ) {
+                        composable(route = ScreenRoutes.SearchPhotosScreen.route) {
+                            // ToDo
+                        }
+                        composable(route = ScreenRoutes.PhotoDetailScreen.route) {
+                            // ToDo
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
