+++ app/src/main/java/com/example/ktunsplashapp/presentation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.ktunsplashapp.ui.theme
+package com.example.ktunsplashapp.presentation.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
